requirements.txt
absl-py==0.13.0
aio-pika==6.8.2
aiofiles==0.8.0
aiohttp==3.7.4
aiormq==3.3.1
APScheduler==3.7.0
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.6.0
beautifulsoup4==4.10.0
bidict==0.21.4
blis==0.7.6
boto3==1.21.8
botocore==1.24.8
Bottleneck==1.3.4
CacheControl==0.12.10
cached-property==1.5.2
cachetools==4.2.4
catalogue==2.0.6
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
clang==5.0
click==7.1.2
cloudpickle==1.6.0
colorclass==2.2.2
coloredlogs==10.0
colorhash==1.0.4
ConfigArgParse==0.13.0
contextvars==2.4
cryptography==3.4.8
cycler==0.11.0
cymem==2.0.6
dataclasses==0.8
decorator==4.4.2
dm-tree==0.1.6
dnspython==1.16.0
docopt==0.6.2
fastai @ https://github.com/fastai/fastai1/archive/master.zip
fastcore==1.3.29
fastdownload==0.0.5
fastprogress==1.0.2
fbmessenger==6.0.0
Flask==1.1.4
Flask-Cors==3.0.10
flatbuffers==1.12
future==0.18.2
gast==0.4.0
gevent==1.5.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.44.0
h11==0.9.0
h5py==3.1.0
httpcore==0.11.1
httptools==0.4.0
httpx==0.15.4
humanfriendly==10.0
idna==3.3
idna-ssl==1.1.0
immutables==0.16
importlib-metadata==4.8.3
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
jsonpickle==2.0.0
jsonschema==3.2.0
kafka-python==2.0.2
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
langcodes==3.3.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.3.4
mattermostwrapper==2.2
msgpack==1.0.3
multidict==5.0.0
murmurhash==1.0.6
networkx==2.5.1
numexpr==2.8.1
numpy==1.19.5
nvidia-ml-py3==7.352.0
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==20.9
pamqp==2.3.0
pandas==1.1.5
pathy==0.6.1
Pillow==8.4.0
preshed==3.0.6
prompt-toolkit==2.0.10
protobuf==3.19.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.8.2
pydot==1.4.2
PyJWT==2.0.0
pykwalify==1.8.0
pymongo==3.10.1
pyparsing==3.0.7
pyrsistent==0.18.0
pyTelegramBotAPI==3.8.3
python-crfsuite==0.9.7
python-dateutil==2.8.2
python-engineio==4.3.1
python-socketio==5.5.2
pytz==2021.3
PyYAML==6.0
questionary==1.9.0
git+https://github.com/guidoivetta/rasa@54d1dac
rasa-sdk==2.8.4
redis==3.5.3
regex==2021.7.6
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
rocketchat-API==1.16.0
rsa==4.8
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.6
s3transfer==0.5.2
sanic==20.9.1
Sanic-Cors==0.10.0.post3
sanic-jwt==1.6.0
Sanic-Plugins-Framework==0.9.5
scikit-learn==0.24.2
scipy==1.5.4
sentry-sdk==1.2.0
six==1.15.0
sklearn-crfsuite==0.3.6
slackclient==2.9.3
smart-open==5.2.1
sniffio==1.2.0
soupsieve==2.3.1
spacy==3.2.3
spacy-legacy==3.0.9
spacy-loggers==1.0.1
SQLAlchemy==1.4.31
srsly==2.4.2
tabulate==0.8.9
tarsafe==0.0.3
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.2
tensorflow-addons==0.14.0
tensorflow-estimator==2.6.0
tensorflow-hub==0.12.0
tensorflow-probability==0.13.0
tensorflow-text==2.6.0
termcolor==1.1.0
terminaltables==3.1.10
thinc==8.0.13
threadpoolctl==3.1.0
toml==0.10.2
torch==1.10.1
torchvision==0.11.2
tqdm==4.62.3
twilio==6.50.1
typeguard==2.13.3
typer==0.4.0
typing==3.7.4.3
typing-extensions==3.7.4.3
tzlocal==2.1
ujson==4.3.0
urllib3==1.26.8
uvloop==0.14.0
wasabi==0.9.0
wcwidth==0.2.5
webexteamssdk==1.6
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.7.2
zipp==3.6.0